requirements.txt
tensorflow==1.14
umap-learn
visdom
librosa>=0.5.1
matplotlib>=2.0.2
numpy>=1.14.0
scipy>=1.0.0
tqdm
sounddevice
SoundFile
Unidecode
inflect
PyQt5
multiprocess
numba==0.48
flask
flask-restplus
pymongo
dnspython
python-dotenv
torch
torchvision
Werkzeug==0.16.1
PyJWT
webrtcvad